+++ field_image_preview/__manifest__.py
@@ -21,7 +21,7 @@ This is extension for <field widget="image"> widget image
 
     # Categories can be used to filter modules in modules listing
     'category': "Tools",
-    'version': '15.1.0.1',
+    'version': '15.1.0.2',
     # any module necessary for this one to work correctly
     'depends': ['web', 'mail'],
     "license": "LGPL-3",