+++ field_image_preview/changelog.rst
@@ -1,1 +1,2 @@
-14.1.0.1 - [FIX] download preview image
+15.1.0.2 - [FIX] print image
+15.1.0.1 - [FIX] download preview image